requirements.txt
easydict==1.9
matplotlib==2.2.3
numpy==1.21.6
Pillow==9.4.0
PyYAML==6.0
torch==1.10.1+cu102
torchvision==0.11.2+cu102
